requirements.txt
alembic==1.13.1
blinker==1.8.2
boto3==1.34.106
botocore==1.34.106
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
urllib3
Werkzeug==3.0.3
